[maven-release-plugin] copy for tag dspace-7.6.3

# Conflicts:
#	.github/workflows/docker.yml
#	dspace-api/pom.xml
#	dspace-api/src/main/java/org/dspace/content/ItemServiceImpl.java
#	dspace-api/src/main/java/org/dspace/content/service/ItemService.java
#	dspace-api/src/main/java/org/dspace/curate/XmlWorkflowCuratorServiceImpl.java
#	dspace-api/src/main/java/org/dspace/discovery/IndexClient.java
#	dspace-api/src/main/java/org/dspace/discovery/SolrServiceImpl.java
#	dspace-api/src/main/java/org/dspace/discovery/indexobject/ItemIndexFactoryImpl.java
#	dspace-api/src/main/java/org/dspace/identifier/doi/DataCiteConnector.java
#	dspace-api/src/main/java/org/dspace/identifier/service/IdentifierService.java
#	dspace-api/src/main/java/org/dspace/orcid/client/OrcidClient.java
#	dspace-api/src/main/java/org/dspace/orcid/client/OrcidClientImpl.java
#	dspace-api/src/main/java/org/dspace/orcid/consumer/OrcidQueueConsumer.java
#	dspace-api/src/main/java/org/dspace/scripts/ProcessServiceImpl.java
#	dspace-api/src/main/java/org/dspace/scripts/service/ProcessService.java
#	dspace-api/src/main/java/org/dspace/versioning/VersioningConsumer.java
#	dspace-api/src/test/data/dspaceFolder/config/local.cfg
#	dspace-api/src/test/java/org/dspace/administer/StructBuilderIT.java
#	dspace-api/src/test/java/org/dspace/builder/AbstractDSpaceObjectBuilder.java
#	dspace-api/src/test/java/org/dspace/builder/WorkspaceItemBuilder.java
#	dspace-api/src/test/java/org/dspace/content/ItemTest.java
#	dspace-api/src/test/java/org/dspace/ctask/general/CreateMissingIdentifiersIT.java
#	dspace-api/src/test/java/org/dspace/orcid/OrcidQueueConsumerIT.java
#	dspace-iiif/pom.xml
#	dspace-oai/pom.xml
#	dspace-rdf/pom.xml
#	dspace-rest/pom.xml
#	dspace-server-webapp/pom.xml
#	dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/RootConverter.java
#	dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/SubmissionSectionConverter.java
#	dspace-server-webapp/src/main/java/org/dspace/app/rest/model/ItemRest.java
#	dspace-server-webapp/src/main/java/org/dspace/app/rest/repository/ProcessRestRepository.java
#	dspace-server-webapp/src/main/java/org/dspace/app/rest/submit/factory/impl/ItemMetadataValueAddPatchOperation.java
#	dspace-server-webapp/src/main/java/org/dspace/app/rest/utils/Utils.java
#	dspace-server-webapp/src/test/java/org/dspace/app/rest/AuthorizationRestRepositoryIT.java
#	dspace-server-webapp/src/test/java/org/dspace/app/rest/BitstreamRestRepositoryIT.java
#	dspace-server-webapp/src/test/java/org/dspace/app/rest/ResearcherProfileRestRepositoryIT.java
#	dspace-server-webapp/src/test/java/org/dspace/app/rest/converter/RootConverterTest.java
#	dspace-services/pom.xml
#	dspace-sword/pom.xml
#	dspace-swordv2/pom.xml
#	dspace/config/crosswalks/DIM2DataCite.xsl
#	dspace/config/dspace.cfg
#	dspace/config/emails/subscriptions_content
#	dspace/config/local.cfg.EXAMPLE
#	dspace/config/log4j2.xml
#	dspace/modules/additions/pom.xml
#	dspace/modules/pom.xml
#	dspace/modules/rest/pom.xml
#	dspace/modules/server/pom.xml
#	dspace/pom.xml
#	pom.xml

+++ dspace-api/src/main/java/org/dspace/orcid/service/impl/OrcidSynchronizationServiceImpl.java
@@ -38,6 +38,7 @@ import org.dspace.eperson.EPerson;
 import org.dspace.eperson.service.EPersonService;
 import org.dspace.orcid.OrcidQueue;
 import org.dspace.orcid.OrcidToken;
+import org.dspace.orcid.client.OrcidClient;
 import org.dspace.orcid.model.OrcidEntityType;
 import org.dspace.orcid.model.OrcidTokenResponseDTO;
 import org.dspace.orcid.service.OrcidQueueService;
@@ -50,6 +51,8 @@ import org.dspace.profile.OrcidProfileSyncPreference;
 import org.dspace.profile.OrcidSynchronizationMode;
 import org.dspace.profile.service.ResearcherProfileService;
 import org.dspace.services.ConfigurationService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 
 /**
@@ -60,6 +63,7 @@ import org.springframework.beans.factory.annotation.Autowired;
  */
 public class OrcidSynchronizationServiceImpl implements OrcidSynchronizationService {
 
+    private static final Logger log = LoggerFactory.getLogger(OrcidSynchronizationServiceImpl.class);
     @Autowired
     private ItemService itemService;
 
@@ -80,6 +84,9 @@ public class OrcidSynchronizationServiceImpl implements OrcidSynchronizationServ
 
     @Autowired
     private ResearcherProfileService researcherProfileService;
+
+    @Autowired
+    private OrcidClient orcidClient;
 
     @Override
     public void linkProfile(Context context, Item profile, OrcidTokenResponseDTO token) throws SQLException {
@@ -128,7 +135,14 @@ public class OrcidSynchronizationServiceImpl implements OrcidSynchronizationServ
         itemService.clearMetadata(context, profile, "dspace", "orcid", "scope", Item.ANY);
         itemService.clearMetadata(context, profile, "dspace", "orcid", "authenticated", Item.ANY);
 
+        OrcidToken profileToken = orcidTokenService.findByProfileItem(context, profile);
+        if (profileToken == null) {
+            log.warn("Cannot find any token related to the user profile: {}", profile.getID());
+            return;
+        }
+
         orcidTokenService.deleteByProfileItem(context, profile);
+        orcidClient.revokeToken(profileToken);
 
         updateItem(context, profile);
